[FIX] conftest: pytest_unconfigure
If a `pytest_configure` crashes, the one from `tmp_path` may not
run (depending the order in which pytest decides to run them), leading
to `config._tmp_path_factory` not being set, leading to our
`pytest_unconfigure` crashing in turn.

Of course it would be a lot better to hook onto the creation of
database templates or something in order to do our cleanup...

+++ conftest.py
@@ -189,9 +189,10 @@ def pytest_unconfigure(config: pytest.Config) -> None:
     if not is_manager(config):
         return
 
-    for c in config._tmp_path_factory.getbasetemp().iterdir():
-        if c.is_file() and c.name.startswith('template-'):
-            subprocess.run(['dropdb', '--if-exists', c.read_text(encoding='utf-8')])
+    if f := getattr(config, '_tmp_path_factory', None):
+        for c in f.getbasetemp().iterdir():
+            if c.is_file() and c.name.startswith('template-'):
+                subprocess.run(['dropdb', '--if-exists', c.read_text(encoding='utf-8')])
 
 @pytest.fixture(scope='session', autouse=True)
 def _set_socket_timeout():
